user is able to register
