+++ routes/authRoute.js
@@ -3,10 +3,12 @@ const passport = require("../middleware/passport");
 const { forwardAuthenticated } = require("../middleware/checkAuth");
 const { json } = require("express");
 const crypto = require("crypto");
+const { registerUser } = require("../controllers/userControllerMongo");
 
 const Joi = require("joi");
 const { ObjectId } = require("bson");
 const router = express.Router();
+const { v4: uuid } = require("uuid");
 
 router.get("/login", forwardAuthenticated, (req, res) => res.render("login"));
 
@@ -52,53 +54,7 @@ router.get("/register", (req, res) => {
 const passwordPepper = "S3cr3+oD3lCli3nt3";
 
 router.post("/register", async (req, res) => {
-  try {
-    console.log("form submit");
-    console.log(req.body);
-
-    // const schema = await Joi.string().max(15).required();
-    // const validationResult = await schema.validate(req.body.first_name);
-    // if (validationResult.error != null) {
-    //   console.log(validationResult.error);
-    //   throw validationResult.error;
-    // }
-
-    const schema = await Joi.object({
-      email: Joi.string().min(5).max(40).required(),
-      password: Joi.string().pattern(new RegExp("^[a-zA-Z0-9]{6,30}$")),
-      username: Joi.string().max(15).required(),
-      email: Joi.string().max(50).required(),
-      //regex allows only letters and number. NOT Special characters
-    });
-    const validationResult = await schema.validate(req.body);
-    if (validationResult.error != null) {
-      console.log(validationResult.error);
-      res.render("error", { message: "Error: Trying to add invalid user" });
-
-      throw validationResult.error;
-    }
-
-    const password_salt = crypto.createHash("sha512");
-
-    password_salt.update(uuid());
-
-    const password_hash = crypto.createHash("sha512");
-
-    password_hash.update(req.body.password + passwordPepper + password_salt);
-
-    const userCollection = database.db("lab_example").collection("users");
-    await userCollection.insertOne({
-      email: req.body.email,
-      username: req.body.username,
-      password_salt: password_salt.digest("hex"),
-      password_hash: password_hash.digest("hex"),
-    });
-    res.redirect("/auth/login");
-  } catch (ex) {
-    res.render("error", { message: "Error connecting to Mongo" });
-    console.log("Error connecting to Mongo");
-    console.log(ex);
-  }
+  registerUser(req, res);
 });
 
 // router.get("/revoke/:sessionID", (req, res) => {
