+++ views/register.ejs
@@ -16,64 +16,72 @@
     <title>Contendr</title>
   </head>
   <body>
-    <nav>
-      <a href="/">
-        <img
-          class="logo"
-          src="/logo-white.png"
-          alt="contendr logo"
-          width="140px"
-          height="auto"
-      /></a>
-    </nav>
-    <h2>challenge your friends!</h2>
-    <form action="/auth/login" method="POST">
-      <div class="form-group">
-        <input
-          type="email"
-          id="email"
-          name="email"
-          class="form-control"
-          placeholder="email"
-        />
-      </div>
+    <div class="wrapper">
+      <nav>
+        <a href="/">
+          <img
+            class="logo"
+            src="/logo-white.png"
+            alt="contendr logo"
+            width="140px"
+            height="auto"
+        /></a>
+      </nav>
       <br />
-      <div class="form-group">
-        <input
-          type="username"
-          id="username"
-          name="username"
-          class="form-control"
-          placeholder="username"
-        />
-      </div>
-      <br />
-      <div class="form-group">
-        <input
-          type="password"
-          id="password"
-          name="password"
-          class="form-control"
-          placeholder="password"
-        />
-      </div>
+      <h2>challenge your friends!</h2>
       <br />
       <br />
-      <br />
-      <button type="submit" class="button" id="register">Register</button>
-    </form>
-    <br />
+      <form action="/auth/register" method="POST">
+        <div class="form-group">
+          <input
+            type="email"
+            id="email"
+            name="email"
+            class="form-control"
+            placeholder="email"
+          />
+        </div>
+        <br />
+        <div class="form-group">
+          <input
+            type="username"
+            id="username"
+            name="username"
+            class="form-control"
+            placeholder="username"
+          />
+        </div>
+        <br />
+        <div class="form-group">
+          <input
+            type="password"
+            id="password"
+            name="password"
+            class="form-control"
+            placeholder="password"
+          />
+        </div>
+        <br />
+        <br />
+        <br />
+        <button type="submit" class="button" id="register">Register</button>
+      </form>
+      <p class="small-text">
+        Already have an account? <a href="/auth/login">Click here to log in.</a>
+      </p>
 
-    <p class="mt-4">
-      Already have an account? <a href="/auth/login">Click here to log in.</a>
-    </p>
-
-    <svg class="wave" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 320">
-      <path
-        fill="#fff"
-        fill-opacity="1"
-        d="M0,160L80,170.7C160,181,320,203,480,181.3C640,160,800,96,960,96C1120,96,1280,160,1360,192L1440,224L1440,320L1360,320C1280,320,1120,320,960,320C800,320,640,320,480,320C320,320,160,320,80,320L0,320Z"
-      ></path>
-    </svg>
+      <svg
+        class="wave"
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 1440 320"
+      >
+        <path
+          fill="#fff"
+          fill-opacity="1"
+          d="M0,160L80,170.7C160,181,320,203,480,181.3C640,160,800,96,960,96C1120,96,1280,160,1360,192L1440,224L1440,320L1360,320C1280,320,1120,320,960,320C800,320,640,320,480,320C320,320,160,320,80,320L0,320Z"
+        ></path>
+      </svg>
+      <footer id="white">.</footer>
+    </div>
   </body>
 </html>
